// management/frontend/src/api/schemas/collectionScenePublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface CollectionScenePublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 64
     */
  name: string;
  /** @maxLength 1000 */
  description?: string;
  /** @maxLength 32 */
  status?: string;
  sort?: number;
  workspace_id: string;
  creator_name?: string | null;
  updater_name?: string | null;
}